requirements.txt
click==8.0.3
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
SQLAlchemy==1.4.27
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.0
mysql-connector-python
